% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_server.R
\name{create_genepanel_server}
\alias{create_genepanel_server}
\title{Create the Shiny server for scGenePanel}
\usage{
create_genepanel_server(data)
}
\arguments{
\item{data}{A Seurat object containing the single-cell data}
}
\value{
A Shiny server function
}
\description{
Create the Shiny server for scGenePanel
}
\keyword{internal}
