% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_ui.R
\name{create_genepanel_ui}
\alias{create_genepanel_ui}
\title{Create the Shiny UI for scGenePanel}
\usage{
create_genepanel_ui(data)
}
\arguments{
\item{data}{A Seurat object used to populate gene choices}
}
\value{
A Shiny UI object
}
\description{
Create the Shiny UI for scGenePanel
}
\keyword{internal}
