% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{discrete_col_palette}
\alias{discrete_col_palette}
\title{Discrete color palettes from multiple packages (varibow, Rcolorbrewer,
ggtheme)}
\usage{
discrete_col_palette(num_colors, hex_cols = NULL, palette = NULL)
}
\arguments{
\item{num_colors}{Number of colors to be generated.}

\item{hex_cols}{hex code of the colors to be selected for violin plot}

\item{palette}{Options are "Tableau" or RColorBrewer qualitative variables
like "Dark2","Paired","Set1","Set2", "Set3", "Accent" etc.
if empty, the function will use based on the entered n.}
}
\value{
A color palette for plotting violin plot
}
\description{
Discrete color palettes from multiple packages (varibow, Rcolorbrewer,
ggtheme)
}
