% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{.check_if_dim_red_exists}
\alias{.check_if_dim_red_exists}
\title{Check if dimension reduction embedding exists}
\usage{
.check_if_dim_red_exists(seurat_obj, dim_red)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{dim_red}{The name of the dimension reduction embedding to check}
}
\value{
returns an error message if the specified dimension reduction does not exist
#' @noRd
}
\description{
Check if the specified dimension reduction embedding exists in the Seurat object
}
