% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{Is_cell_type_name}
\alias{Is_cell_type_name}
\title{Check if cell type name exists}
\usage{
Is_cell_type_name(seurat_object, cell_type_name, cell_type_colname)
}
\arguments{
\item{seurat_object}{Seurat object name.}

\item{cell_type_name}{cell type name to highlight}

\item{cell_type_colname}{The metadata column name for the cell identity annotations}
}
\value{
returns an error message if cell type name does not exist
}
\description{
Check to see if cell type  name exist in the seurat object, if not returns an error
}
