% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{Is_celltype_colname}
\alias{Is_celltype_colname}
\title{Check if cell type metadata exists}
\usage{
Is_celltype_colname(seurat_object, cell_type_colname)
}
\arguments{
\item{seurat_object}{Seurat object name.}

\item{cell_type_colname}{cell type column name in metadata}
}
\value{
returns an error message if cell_type_colname does not exist
}
\description{
Check to see if cell type metadata column name exist in the seurat object, if not returns an error
}
