% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{Is_gene}
\alias{Is_gene}
\title{Check gene exists}
\usage{
Is_gene(seurat_object, gene)
}
\arguments{
\item{seurat_object}{Seurat object name.}

\item{gene}{name of gene}
}
\value{
returns an error message if gene is not found in seurat object
}
\description{
Check if gene included in object
}
