% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{Is_meta_group}
\alias{Is_meta_group}
\title{Check meta_group}
\usage{
Is_meta_group(seurat_object, meta_group)
}
\arguments{
\item{seurat_object}{Seurat object name.}

\item{meta_group}{name of group to split visuals by}
}
\value{
returns an error message if meta_group does not exist
}
\description{
Check if 'meta_group' exists
}
