% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{Is_Seurat}
\alias{Is_Seurat}
\title{Check Seurat Object}
\usage{
Is_Seurat(seurat_object)
}
\arguments{
\item{seurat_object}{Seurat object name.}
}
\value{
returns an error message if the object is not seurat
}
\description{
Checks if object is a Seurat object
}
