% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{make_seurat}
\alias{make_seurat}
\title{Check Seurat Object}
\usage{
make_seurat(object)
}
\arguments{
\item{object}{A Seurat or SingleCellExperiment object}
}
\value{
returns an error message if the object is not among the expected
input object
}
\description{
Convert SingleCellExperiment object to Seurat object
}
