% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGenePanel-package.R
\docType{package}
\name{scGenePanel-package}
\alias{scGenePanel}
\alias{scGenePanel-package}
\title{scGenePanel: Multi-panel visualization for single cell expression data}
\description{
A package for creating a multi-panel view of a single gene within a selected cell population and variable trait, including UMAP embeddings, violin plots and expression metrics derived from single cell (or nucleus) gene expression data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vandydata/scGenePanel}
  \item Report bugs at \url{https://github.com/vandydata/scGenePanel/issues}
}

}
\author{
\strong{Maintainer}: Shristi Shrestha \email{shristi.shrestha.2@vanderbilt.edu} (\href{https://orcid.org/0000-0003-4519-1712}{ORCID})

Authors:
\itemize{
  \item Jean-Philippe Cartailler \email{jp.cartailler@vanderbilt.edu} (\href{https://orcid.org/0000-0002-0312-2391}{ORCID})
}

}
\keyword{internal}
