% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_panel.R
\name{umap_panel}
\alias{umap_panel}
\title{This functions generates UMAP plots for scGenePanel}
\usage{
umap_panel(
  seurat_obj,
  cell_type_colname,
  cell_type_name,
  meta_group,
  gene,
  levels_idents,
  group_order = NULL
)
}
\arguments{
\item{cell_type_colname}{The metadata column name that contains the cell
identity annotations}

\item{cell_type_name}{The cell type identity to highlight in UMAP}

\item{meta_group}{The metadata column name of the variable to split the UMAP,
violinplot and cell frequency table by. for example, to split by disease
condition}

\item{gene}{Name of gene to explore gene expression}

\item{levels_idents}{The levels of the metadata column to split with}

\item{group_order}{User defined order of the meta_groups to be displayed}

\item{object}{A Seurat object}
}
\value{
ggplot2 UMAP plot objects
}
\description{
This functions generates UMAP plots for scGenePanel
}
