% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violinpanel.R
\name{violin_panel}
\alias{violin_panel}
\title{scRNAseq multipanel gene expression visual}
\usage{
violin_panel(
  object,
  cell_type_colname,
  cell_type_name,
  meta_group,
  gene,
  col.palette
)
}
\arguments{
\item{object}{A Seurat or SingleCellExperiment}

\item{cell_type_colname}{The metadata column name that contains the cell identity annotations}

\item{cell_type_name}{The cell type identity to highlight}

\item{meta_group}{The metadata column name that contains the groups identity per cell}

\item{gene}{Name of gene to explore gene expression in UMAP, violin plot, and cell frequency table}

\item{col.palette}{Color palettes to choose for violinplot panel. Options are "tableu","varibow" or RColorBrewer qualitative variables like "Dark2","Paired","Set1" etc}
}
\value{
Multipanel plots
}
\description{
This functions exports a multipanel scRNAseq gene expression plots of UMAP, violin plot with usen defined cell type and condition/groups
along with tabular cell counts and ratios per chosen group in one visual
}
