% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellfreq_panel.R
\name{cellfreq_panel}
\alias{cellfreq_panel}
\title{This functions generates cell frequency table for scGenePanel}
\usage{
cellfreq_panel(
  seurat_obj,
  cell_type_colname,
  cell_type_name,
  meta_group,
  gene,
  col_palette,
  levels_idents
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{cell_type_colname}{The metadata column name that contains the cell
identity annotations}

\item{cell_type_name}{The cell type identity to highlight}

\item{meta_group}{The metadata column name of the variable to split the UMAP,
violinplot and cell frequency table by. for example, to split by disease
condition}

\item{gene}{Name of gene to explore gene expression in UMAP, violin plot and
cell frequency table}

\item{col_palette}{Color palettes to choose for violinplot panel. Options
are "tableu","varibow" or RColorBrewer qualitative variables like "Dark2",
"Paired", "Set1" etc}

\item{levels_idents}{The levels of the metadata column to split with}
}
\value{
ggpubr table
}
\description{
This functions generates cell frequency table for scGenePanel
}
