% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepanel.R
\name{create_gene_panel}
\alias{create_gene_panel}
\title{scRNAseq multipanel gene expression visual by sample groups}
\usage{
create_gene_panel(
  object,
  gene,
  meta_group,
  cell_type_name,
  cell_type_colname,
  col.palette,
  group_order = NULL,
  output_dir = "."
)
}
\arguments{
\item{object}{A Seurat or SingleCellExperiment}

\item{gene}{Name of gene to explore gene expression in UMAP, violin plot, and cell frequency table}

\item{meta_group}{The metadata column name that contains the groups identity per cell}

\item{cell_type_name}{The cell type identity to highlight}

\item{cell_type_colname}{The metadata column name that contains the cell identity annotations}

\item{col.palette}{Color palettes to choose for violinplot panel. Options are "tableu","varibow" or RColorBrewer qualitative variables like "Dark2","Paired","Set1" etc}

\item{group_order}{User defined order of the groups to be displayed}

\item{output_dir}{Output directory where the image will be saved}
}
\value{
Multipanel plots
}
\description{
\code{create_gene_panel} exports a multipanel scRNAseq gene expression plots of UMAPs clusters highlighted with user defined cell types and split view on user defined sample groups, violin plot with user defined sample groups
along with tabular cell counts, ratios and expression values per sample group, in one visual
}
