% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{Is_cell_type_name}
\alias{Is_cell_type_name}
\title{Check if cell type name exists}
\usage{
Is_cell_type_name(object, cell_type_name, cell_type_colname)
}
\arguments{
\item{object}{A Seurat or SingleCellExperiment object}

\item{cell_type_name}{The cell type identity to highlight in UMAP}

\item{cell_type_colname}{The metadata column name that contains the cell identity annotations}
}
\value{
returns an error message if cell type name does not exist
}
\description{
Check to see if cell type  name exist in the seurat object, if not returns an error
}
