% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{Is_celltype_colname}
\alias{Is_celltype_colname}
\title{Check if cell type metadata exists}
\usage{
Is_celltype_colname(object, cell_type_colname)
}
\arguments{
\item{object}{A Seurat or SingleCellExperiment object}

\item{cell_type_colname}{The metadata column name that contains the cell identity annotations}
}
\value{
returns an error message if cell_type_colname does not exist
}
\description{
Check to see if cell type metadata column name exist in the seurat object, if not returns an error
}
