% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umappanel.R
\name{umap_panel}
\alias{umap_panel}
\title{scRNAseq multi-panel gene expression visual}
\usage{
umap_panel(
  object,
  cell_type_colname,
  cell_type_name,
  meta_group,
  gene,
  col_palette
)
}
\arguments{
\item{object}{A Seurat or SingleCellExperiment}

\item{cell_type_colname}{The metadata column name that contains the cell
identity annotations}

\item{cell_type_name}{The cell type identity to highlight in UMAP}

\item{meta_group}{The metadata column name of the variable to split the UMAP,
violinplot and cell frequency table by. for example, to split by disease
condition}

\item{gene}{Name of gene to explore gene expression in UMAP, violin plot, and
cell frequency table}

\item{col_palette}{Color palettes to choose for violinplot panel. Options are
"tableu" or RColorBrewer qualitative variables like "Dark2", "Paired",
"Set1", "Set2", "Set3", "Accent" etc.}
}
\value{
multi-panel plots
}
\description{
This functions exports a multi-panel scRNAseq gene expression plots of UMAP,
violin plot with usen defined cell type and condition/groups along with
tabular cell counts and ratios per chosen group in one visual
}
