% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_violin.R
\name{violin_panel}
\alias{violin_panel}
\title{This functions generates violin plots for scGenePanel}
\usage{
violin_panel(
  seurat_obj,
  cell_type_colname,
  cell_type_name,
  meta_group,
  gene,
  col_palette,
  levels_idents
)
}
\arguments{
\item{cell_type_colname}{The metadata column name that contains the cell
identity annotations}

\item{cell_type_name}{The cell type identity to highlight}

\item{meta_group}{The metadata column name of the variable to split with}

\item{gene}{Name of gene to explore gene expression}

\item{col_palette}{Color palettes to choose for violinplot panel. Options are
"tableu" or RColorBrewer qualitative variables like "Dark2", "Paired",
"Set1", "Set2", "Set3", "Accent" etc.}

\item{object}{A Seurat object}
}
\value{
ggplot2 violing plot object
}
\description{
This functions generates violin plots for scGenePanel
}
